Move plugin registry into legacy folder to prep for new implementation
