+++ ui/plugin-system/src/components/PluginRegistry/legacy/PluginRegistry.tsx
