+++ ui/plugin-system/src/components/PluginRegistry/legacy/index.ts
