+++ ui/plugin-system/src/components/PluginRegistry/legacy/registry-state.ts
