+++ ui/plugin-system/src/components/index.ts
@@ -13,4 +13,4 @@
 
 export * from './PluginBoundary';
 export * from './PluginLoadingBoundary';
-export * from './PluginRegistry';
+export * from './PluginRegistry/legacy';
